support #23 #21: pac support and fix config update
