+++ conf/api.config
@@ -1,7 +1,5 @@
 listenaddr: 0.0.0.0:8888
 debugmode: true
-mysqldsn: api:api@tcp(10.0.0.2:3306)/api?charset=utf8mb4&parseTime=True&loc=Local
-redisdsn: ""
 pachandlerconfig:
   pacgeneratecron: 0 0 * * *
-  pacproxyaddr: SOCKS5 10.0.0.2:1081
+  pacgeneratecmd: /usr/local/bin/genpac --format pac --gfwlist-proxy 'SOCKS5 10.0.0.2:1081' --pac-proxy 'SOCKS5 10.0.0.2:1081' --user-rule '%v'
