+++ conf/v2ray.json
@@ -1,1 +1,122 @@
-{"inbounds":[{"port":1081,"listen":"0.0.0.0","protocol":"socks","settings":{"udp":true}},{"port":1082,"protocol":"dokodemo-door","settings":{"network":"tcp,udp","followRedirect":true,"timeout":30}},{"port":1083,"protocol":"http","allowTransparent":false,"userLevel":0}],"log":{"error":"v2ray_error.log","loglevel":"warning","access":"v2ray_access.log"},"outbounds":[{"sendThrough":"0.0.0.0","mux":{"enabled":false,"concurrency":8},"protocol":"vmess","settings":{"vnext":[{"address":"{serverHost}","users":[{"id":"{serverId}","alterId":2,"security":"auto","level":0}],"port":9495945}]},"tag":"{serverName}","streamSettings":{"wsSettings":{"path":"{serverPath}","headers":{"Host":"{serverHost}"}},"quicSettings":{"key":"","header":{"type":"none"},"security":"none"},"tlsSettings":{"allowInsecure":false,"alpn":["http/1.1"],"serverName":"{serverHost}","allowInsecureCiphers":false},"sockopt":{},"httpSettings":{"path":"","host":[""]},"tcpSettings":{"header":{"type":"none"}},"kcpSettings":{"header":{"type":"none"},"mtu":1350,"congestion":false,"tti":20,"uplinkCapacity":5,"writeBufferSize":1,"readBufferSize":1,"downlinkCapacity":20},"security":"tls","network":"ws"}}],"routing":{"domainStrategy":"IPOnDemand","rules":[{"type":"field","ip":["geoip:private"],"outboundTag":"direct"}]}}
+{
+    "inbounds":[
+        {
+            "port":1081,
+            "listen":"0.0.0.0",
+            "protocol":"socks",
+            "settings":{
+                "udp":true
+            }
+        },
+        {
+            "port":1082,
+            "protocol":"dokodemo-door",
+            "settings":{
+                "network":"tcp,udp",
+                "followRedirect":true,
+                "timeout":30
+            }
+        },
+        {
+            "port":1083,
+            "protocol":"http",
+            "allowTransparent":false,
+            "userLevel":0
+        }
+    ],
+    "log":{
+        "error":"v2ray_error.log",
+        "loglevel":"warning",
+        "access":"v2ray_access.log"
+    },
+    "outbounds":[
+        {
+            "sendThrough":"0.0.0.0",
+            "mux":{
+                "enabled":false,
+                "concurrency":8
+            },
+            "protocol":"vmess",
+            "settings":{
+                "vnext":[
+                    {
+                        "address":"{serverHost}",
+                        "users":[
+                            {
+                                "id":"{serverId}",
+                                "alterId":2,
+                                "security":"auto",
+                                "level":0
+                            }
+                        ],
+                        "port":9495945
+                    }
+                ]
+            },
+            "tag":"{serverName}",
+            "streamSettings":{
+                "wsSettings":{
+                    "path":"{serverPath}",
+                    "headers":{
+                        "Host":"{serverHost}"
+                    }
+                },
+                "quicSettings":{
+                    "key":"",
+                    "header":{
+                        "type":"none"
+                    },
+                    "security":"none"
+                },
+                "tlsSettings":{
+                    "allowInsecure":false,
+                    "alpn":[
+                        "http/1.1"
+                    ],
+                    "serverName":"{serverHost}",
+                    "allowInsecureCiphers":false
+                },
+                "sockopt":{
+
+                },
+                "httpSettings":{
+                    "path":"",
+                    "host":[
+                        ""
+                    ]
+                },
+                "tcpSettings":{
+                    "header":{
+                        "type":"none"
+                    }
+                },
+                "kcpSettings":{
+                    "header":{
+                        "type":"none"
+                    },
+                    "mtu":1350,
+                    "congestion":false,
+                    "tti":20,
+                    "uplinkCapacity":5,
+                    "writeBufferSize":1,
+                    "readBufferSize":1,
+                    "downlinkCapacity":20
+                },
+                "security":"tls",
+                "network":"ws"
+            }
+        }
+    ],
+    "routing":{
+        "domainStrategy":"IPOnDemand",
+        "rules":[
+            {
+                "type":"field",
+                "ip":[
+                    "geoip:private"
+                ],
+                "outboundTag":"direct"
+            }
+        ]
+    }
+}
