+++ front-end/package.json
@@ -8,13 +8,15 @@
     "@testing-library/user-event": "^12.1.10",
     "antd": "^4.7.3",
     "axios": "^0.21.0",
-    "codemirror": "^5.58.2",
+    "codemirror": "^5.59.4",
+    "re-resizable": "^6.9.0",
     "react": "^17.0.1",
     "react-codemirror2": "^7.2.1",
     "react-dom": "^17.0.1",
     "react-draggable": "^4.4.3",
     "react-json-editor-ajrm": "^2.5.13",
     "react-json-view": "^1.19.1",
+    "react-monaco-editor": "^0.42.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.0",
     "web-vitals": "^0.2.4"
