+++ handler/handler.go
@@ -1,63 +1,24 @@
 package handler
 
 import (
-	"fmt"
-	"net/http"
-
-	"github.com/cntechpower/v2ray-webui/model"
-	"github.com/gin-gonic/gin"
-	"github.com/go-playground/validator/v10"
+	"github.com/cntechpower/v2ray-webui/handler/file"
+	"github.com/cntechpower/v2ray-webui/handler/pac"
+	"github.com/cntechpower/v2ray-webui/handler/v2ray"
 )
 
-var checker *validator.Validate
-
-func init() {
-	checker = validator.New()
-}
-
-var Pac *PacHandler
-var Status *statusHandler
-var V2ray *v2rayHandler
-var File *fileHandler
+var Pac *pac.Handler
+var V2ray *v2ray.Handler
+var File *file.Handler
 
 func Init(v2rayTemplateConfigFilePath string) (err error) {
-	Status = newStatusHandler()
-	Pac, err = newPacHandler()
+	Pac, err = pac.New()
 	if err != nil {
 		return
 	}
-	V2ray, err = newV2rayHandler(v2rayTemplateConfigFilePath)
+	V2ray, err = v2ray.New(v2rayTemplateConfigFilePath)
 	if err != nil {
 		return
 	}
-	File = newFileHandler()
+	File = file.New()
 	return
-}
-
-type baseHandler struct {
-}
-
-func (h *baseHandler) DoFunc(c *gin.Context, f func() error) {
-	if err := f(); err != nil {
-		c.AbortWithStatusJSON(http.StatusInternalServerError, model.NewGenericStatus(http.StatusInternalServerError, err.Error()))
-		return
-	}
-	c.JSON(http.StatusOK, model.NewGenericStatus(http.StatusOK, "Operation Succeed."))
-}
-
-func (h *baseHandler) GenericWrapper(f func() error) func(c *gin.Context) {
-	return func(c *gin.Context) {
-		if err := f(); err != nil {
-			c.AbortWithStatusJSON(http.StatusInternalServerError, model.NewGenericStatus(http.StatusInternalServerError, err.Error()))
-			return
-		}
-		c.JSON(http.StatusOK, model.NewGenericStatus(http.StatusOK, "Operation Succeed."))
-	}
-
-}
-func errorWith500(c *gin.Context, err error) {
-	c.AbortWithStatusJSON(http.StatusInternalServerError, model.NewGenericStatus(http.StatusInternalServerError, err.Error()))
-}
-func ok(c *gin.Context, message string, a ...interface{}) {
-	c.JSON(http.StatusOK, model.NewGenericStatus(http.StatusOK, fmt.Sprintf(message, a...)))
 }
