+++ handler/pac/pac_test.go
@@ -1,4 +1,4 @@
-package handler
+package pac
 
 import (
 	"testing"
