--- handler/status.go
@@ -1,11 +1,0 @@
-package handler
-
-type statusHandler struct {
-	*baseHandler
-}
-
-func newStatusHandler() *statusHandler {
-	return &statusHandler{
-		&baseHandler{},
-	}
-}
